Fabric Crypto Service Provider
This change-set introduces the interfaces to define a Crypto Service Provider.
The CSP will provides the implementation of cryptographic standards and
algorithms, and it will be pluggable.
This change-set represents the first step in the context of:
https://jira.hyperledger.org/browse/FAB-354
Subsequents change-sets will refactor the membership-services and the core
to use the CSP only.

Change-Id: Idf4677d41b3111b6971b37f94c8e2f014b1af0a9
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Volodymyr Paprotski <vpaprots@ca.ibm.com>