Default Fabric Crypto Service Provider
This change-set introduces a default software-based implementation for the
Crypto Service Provider introduced by https://gerrit.hyperledger.org/r/#/c/1605/
The implementation is based on the 'primitives' package and provides only
ecdsa key generation, sign, verify and public key marshalling.
In addition this change-set introduced an enhachment for the Key interface
definition.
This change-set refers to the following JIRA:
https://jira.hyperledger.org/browse/FAB-354
This change-set should be merged after
https://gerrit.hyperledger.org/r/#/c/1605/.

Change-Id: Ie43711633db76e52bbb0382b8d40e41e4621317c
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Volodymyr Paprotski <vpaprots@ca.ibm.com>