Key Re-randomization Support for Fabric CSP
This change-set introduces ECDSA key re-randomization as required by TCert
generation. In addition, simplification has been introduced in the key-store
and its configuration.
In the subsequent change-sets we will start using the CSP at the core and
membership service, at least for all the required signing-related operations.
This change-set refers to the following JIRA:
https://jira.hyperledger.org/browse/FAB-354
This change-set should be merged after
https://gerrit.hyperledger.org/r/#/c/1645/

Change-Id: If28eac4a4619392bbe6ae8ba1521cd30966d88be
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Volodymyr Paprotski <vpaprots@ca.ibm.com>