BCCSP Factory support
This change-set introduces support for multiple BCCSP factories.
This way, it will be possible to have software-based as well as
hardware-based BCCSP implementations.
The concept of default BCCSP still remains and represents the BCCSP
to be used by default as it is configured via viper.
Tests have been added to test the factories.
Some name refactoring has been performed to better identify the factories.
This change-set represents the first step in the context of:
https://jira.hyperledger.org/browse/FAB-354
This change-set should be merged after
https://gerrit.hyperledger.org/r/#/c/1695/

Change-Id: I806dd24abce991062f54f904f64efb3b5bda9bc2
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>
Signed-off-by: Volodymyr Paprotski <vpaprots@ca.ibm.com>